requirements.txt
streamlit
psycopg2-binary
faker
pandas
